refactor: controller
